Some of Block-memory interface complete

+++ inode.hpp
@@ -12,38 +12,38 @@ namespace ModV6FileSystem
         struct Data
         {
         public:
-            unsigned short flags;
-            unsigned short nlinks;
-            unsigned int uid;
-            unsigned int gid;
+            uint16_t flags;
+            uint16_t nlinks;
+            uint32_t uid;
+            uint32_t gid;
             // assuming this is earlier bytes
-            unsigned int size1;
+            uint32_t size1;
             // assuming this is later bytes
-            unsigned int size2;
-            std::array<unsigned int, 9> addr;
-            unsigned int actime;
-            unsigned int modtime;
+            uint32_t size2;
+            std::array<uint32_t, 9> addr;
+            uint32_t actime;
+            uint32_t modtime;
         };
         Data _data;
 
     public:
         friend std::ostream &operator<<(std::ostream &ostream, const INode& in);
         ~INode();
-        unsigned short flags() const;
-        void flags(unsigned short flags);
-        unsigned short nlinks() const;
-        void nlinks(unsigned short nlinks);
-        unsigned int uid() const;
-        void uid(unsigned int uid);
-        unsigned int gid() const;
-        void gid(unsigned int gid);
-        unsigned long int size() const;
-        void size(unsigned long long size);
-        std::array<unsigned int, 9> addr() const;
-        void addr(std::array<unsigned int, 9> addr);
-        unsigned int actime() const;
-        void actime(unsigned int actime);
-        unsigned int modtime() const;
-        void modtime(unsigned int modtime);
+        uint16_t flags() const;
+        void flags(uint16_t flags);
+        uint16_t nlinks() const;
+        void nlinks(uint16_t nlinks);
+        uint32_t uid() const;
+        void uid(uint32_t uid);
+        uint32_t gid() const;
+        void gid(uint32_t gid);
+        uint64_t size() const;
+        void size(uint64_t size);
+        std::array<uint32_t, 9> addr() const;
+        void addr(std::array<uint32_t, 9> addr);
+        uint32_t actime() const;
+        void actime(uint32_t actime);
+        uint32_t modtime() const;
+        void modtime(uint32_t modtime);
     };
 }
